replace double with float for performance on F407 and the like (cortexM4F)

+++ cppext/argset.h
@@ -5,21 +5,21 @@
 #include "buffer.h"
 
 /** used as an intermediate representation of a struct made of only numbers (no text) */
-class ArgSet : public Indexer<double> {
+class ArgSet : public Indexer<float> {
 public:
-/** wrap an existing array of doubles */
-  ArgSet(double *d, unsigned sizeofd);
+/** wrap an existing array of floats */
+  ArgSet(float *d, unsigned sizeofd);
 
   ArgSet(const ArgSet &other, int portion);
 
 /** @return whether @param changed*/
-  bool applyto(double&d);
+  bool applyto(float&d);
 /** @return whether @param changed*/
   bool applyto(int&eye);
 /** @return whether @param changed*/
   bool applyto(bool&b);
 /** append an array of const data, can't just wrap them without lots of syntax*/
-  void cat(const double *prefilled,int qty);
+  void cat(const float *prefilled,int qty);
 /** @returns whether this and @param args are same size, and whether corresponding entries are nearly equal.
  *  @see nearly for @param bits value */
   bool equals(const ArgSet &args, int bits = 18) const;
@@ -30,14 +30,14 @@ public:
 
 /** replacing macro with real class */
 template<unsigned qty> class ArgBlock : public ArgSet {
-  double argv[qty];
+  float argv[qty];
 
 public:
   ArgBlock() : ArgSet(argv,sizeof(argv)){
   }
 
   /** @returns a reference to a value, the first if arg is bad. */
-  double &operator [](unsigned which){
+  float &operator [](unsigned which){
     return argv[which<qty ? which : 0];
   }
 
@@ -47,10 +47,10 @@ public:
 
 }; // class ArgBlock
 
-/** exists so that we can wrap const arrays of doubles for firmware rom based config. */
-class ConstArgSet : public Indexer<const double> {
+/** exists so that we can wrap const arrays of floats for firmware rom based config. */
+class ConstArgSet : public Indexer<const float> {
 public:
-  ConstArgSet(const double *d, unsigned numArgs);
+  ConstArgSet(const float *d, unsigned numArgs);
   ConstArgSet(const ArgSet &other);
   ConstArgSet(const ConstArgSet &other);
   ~ConstArgSet() = default;
@@ -58,7 +58,7 @@ public:
 
 //FakeArgs takes a root name and creates a const array and a wrapper for it: nameArgs
 #define FakeArgs(name, ...)\
-static constexpr double name##Scaling[] = { __VA_ARGS__ };\
+static constexpr float name##Scaling[] = { __VA_ARGS__ };\
 static ConstArgSet name##Args(&name##Scaling[0], sizeof(name##Scaling))
